# Remove block records table