The rationale for this patch is to simplify the database and make it smaller. The key (`header_hash`) in both `block_records` and `full_blocks` tables are 32 bytes, and by storing the block_records in the same table as the full blocks, we only need a single copy of those keys.

The first version of this patch removed the serialization of the BlockRecord object entirely, and rebuilt it, on demand, from the FullBlock object. However, until we fix parsing the FullBlock performance issue, that takes too long. It made the benchmark take a lot longer.

This patch, instead, stores the BlockRecord in the same table as the full block, meaning it can use the same header_hash key as the full block.

It's hard to explain why this makes the table slightly larger. If anything it should make it smaller. Perhaps there's some sqlite implementation detail or a scale issue (perhaps the difference is only observed at a larger scale than the benchmark).

Note that the main contributor to slowing down the benchmarks is the `get_random_not_compactified` test. I have another patch to improve this, by adding a column to indicate whether a block is part of the main chain or not.

test  | before | after | after / before
---|---|---|---
table size | 352.5 MB | 357.4 MB | 101.4%
Ryzen, fast SSD | 308.8 s | 317.5 s | 102.8%
RPi, SD card | 1614.2s | 1381.3s | 116.9 %

# raw benchmark output

this patch:

Ryzen, fast SSD:

```
version 2
46.2831s, add_full_block
52.6006s, get_full_block
5.3804s, get_full_block_bytes
51.5568s, get_full_blocks_at
6.3655s, get_block_records_by_hash
52.9810s, get_blocks_by_hash
5.6967s, get_block_record
0.6920s, get_block_records_in_range
0.0064s, get_block_records_close_to_peak
4.2865s, get_block_record
91.6701s, get_random_not_compactified
all tests completed in 317.5190s
database size: 357.372 MB
```

RPi, SD card:

```
version 2
262.1600s, add_full_block
202.8422s, get_full_block
23.8295s, get_full_block_bytes
198.1242s, get_full_blocks_at
26.1504s, get_block_records_by_hash
203.2471s, get_blocks_by_hash
25.6698s, get_block_record
3.4417s, get_block_records_in_range
0.0343s, get_block_records_close_to_peak
17.4574s, get_block_record
651.2877s, get_random_not_compactified
all tests completed in 1614.2444s
database size: 357.372 MB
```

previous commit:

Ryzen, fast SSD:

```
version 2
56.3806s, add_full_block
53.0442s, get_full_block
5.1803s, get_full_block_bytes
50.1041s, get_full_blocks_at
6.8339s, get_block_records_by_hash
52.6259s, get_blocks_by_hash
6.3245s, get_block_record
0.6537s, get_block_records_in_range
0.0056s, get_block_records_close_to_peak
4.9460s, get_block_record
72.6523s, get_random_not_compactified
all tests completed in 308.7509s
database size: 352.473 MB
```

RPi, SD card:

```
version 2
305.0027s, add_full_block
204.0082s, get_full_block
24.1904s, get_full_block_bytes
198.7703s, get_full_blocks_at
24.1730s, get_block_records_by_hash
204.2618s, get_blocks_by_hash
24.6884s, get_block_record
2.8021s, get_block_records_in_range
0.0261s, get_block_records_close_to_peak
18.0473s, get_block_record
375.3662s, get_random_not_compactified
all tests completed in 1381.3366s
database size: 352.473 MB

```