@arvidn 2021-12-15
> Is it difficult to run the benchmarks with larger scenarios to test the scaling theory?

Not very difficult, it's quite slow (generating block is the thing that take the longest). I'll rerun the benchmark with a 10x test size.