@arvidn 2021-12-15
even when using a larger test, the new database is slightly larger than the previous one. (this is looking at the size of the .sqlite file)

3527 MB -> 3591 MB

Querying the DB for the table sizes:

this commit:

```
sqlite> SELECT SUM("pgsize") FROM "dbstat" WHERE name='full_blocks';
3577200640
```

prior commit:

```
sqlite> SELECT SUM("pgsize") FROM "dbstat" WHERE name='full_blocks';
3364847616
sqlite> SELECT SUM("pgsize") FROM "dbstat" WHERE name='block_records';
137052160
```

Sum: 3501899776

Which is smaller than the combined table.
 I'll look closer.