@arvidn 2021-12-15
running `py-sqlite-analyzer` on these databases, one interesting observation is:

before:

```
Unused bytes on index pages .................................. 165227    15.05%
Unused bytes on primary pages .............................. 26656930     7.64%
Unused bytes on overflow pages ............................... 185922     0.01%
Unused bytes on all pages .................................. 27008079     0.77%
```

after:

```
Unused bytes on index pages .................................. 155730    13.63%
Unused bytes on primary pages ............................. 109475366    26.54%
Unused bytes on overflow pages .................................... 0     0.00%
Unused bytes on all pages ................................. 109631096     3.05%
```

That's 109 MB unused versus 27 MB. So, it seems to be some kind of fragmentation perhaps.