@altendky 2021-12-15
Maybe a vacuum makes them look more as expected?