@arvidn 2021-12-15
only marginally different:

```
Unused bytes on index pages .................................. 143468    12.83%
Unused bytes on primary pages ............................. 107960212    26.27%
Unused bytes on overflow pages .................................... 0     0.00%
Unused bytes on all pages ................................. 108103680     3.01%
```